# foundit-rest-api/app/api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "Foundit REST API",
    "description": "This is the REST API backend for Foundit."
  },
  "host": "localhost:5000",
  "basePath": "/api",
  "schemes": [
    "http"
  ],
  "paths": {
    "/items/": {
      "get": {
        "tags": [
          "Items"
        ],
        "description": "Get all items",
        "parameters": [
          {
            "name": "page",
            "description": "Page number",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Invalid page number"
          }
        }
      },
      "post": {
        "tags": [
          "Items"
        ],
        "description": "Add an item",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "imgUrl": {
                  "example": "any"
                },
                "category": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Missing required fields"
          }
        }
      }
    },
    "/items/{id}": {
      "get": {
        "tags": [
          "Items"
        ],
        "description": "Get an item by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Missing required fields"
          },
          "404": {
            "description": "Item not found"
          }
        }
      }
    },
    "/": {
      "get": {
        "tags": [
          "API Information"
        ],
        "description": "Information about the API",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  }
}